# Optimize EnsureCapacity