--- a/src/LinkDotNet.StringBuilder/ValueStringBuilder.Insert.cs
+++ b/src/LinkDotNet.StringBuilder/ValueStringBuilder.Insert.cs
@@ -45,7 +45,7 @@ public ref partial struct ValueStringBuilder
         var newLength = bufferPosition + value.Length;
         if (newLength > buffer.Length)
         {
-            Grow(newLength);
+            EnsureCapacity(newLength);
         }
 
         bufferPosition = newLength;
@@ -79,7 +79,7 @@ public ref partial struct ValueStringBuilder
             var newLength = bufferPosition + written;
             if (newLength > buffer.Length)
             {
-                Grow(newLength);
+                EnsureCapacity(newLength);
             }
 
             bufferPosition = newLength;