--- a/tests/LinkDotNet.StringBuilder.UnitTests/ValueStringBuilderExtensionsTests.cs
+++ b/tests/LinkDotNet.StringBuilder.UnitTests/ValueStringBuilderExtensionsTests.cs
@@ -31,7 +31,7 @@ public class ValueStringBuilderExtensionsTests
     {
         System.Text.StringBuilder? sb = null;
 
-        Action act = () => sb.ToValueStringBuilder();
+        Action act = () => sb!.ToValueStringBuilder();
 
         act.ShouldThrow<ArgumentNullException>();
     }