@Joy-less 2025-01-24
This is a very small optimization if at all I think but it's simpler so that's good.